extract shared AppShell layout for all protected pages
Refactors the sidebar and header out of individual page components into a single shared AppShell component.

- New AppShell component with role-filtered nav, user avatar, and sign-out button
- _protected.tsx route wraps Outlet with AppShell — all current and future protected pages get the sidebar automatically
- Dashboard, AppointmentsPage, and AppointmentDetailsPage stripped of duplicated shell code
- Fixes sidebar active-item hover bug where text turned black on the currently active nav item

Files to stage:
  frontend/src/components/layout/AppShell.tsx  (new)
  frontend/src/routes/_protected.tsx
  frontend/src/features/dashboard/Dashboard.tsx
  frontend/src/features/appointments/AppointmentsPage.tsx
  frontend/src/features/appointments/AppointmentDetailsPage.tsx
